requirements.txt
paper-qa==2026.3.18
pdfplumber
python-dotenv==1.2.2
streamlit==1.56.0
